refactor: move algorithm lists to their own file
